Core: include optimization (glm)
