+++ include/inviwo/core/properties/property.h
@@ -41,7 +41,6 @@
 #include <inviwo/core/util/document.h>
 #include <inviwo/core/metadata/metadataowner.h>
 #include <inviwo/core/util/introspection.h>
-#include <inviwo/core/io/serialization/serialization.h>
 
 #include <functional>
 #include <type_traits>
@@ -52,6 +51,8 @@
 namespace inviwo {
 
 class NetworkVisitor;
+class Serializer;
+class Deserializer;
 
 /**
  * \class PropertyTraits
