+++ include/inviwo/core/util/glmutils.h
@@ -208,6 +208,32 @@ constexpr auto glmcomp(const T& elem, size_t i, size_t j) -> const typename T::v
     return elem[i][j];
 }
 
+template <typename T = double, glm::length_t C = 1, glm::length_t R = 1,
+          glm::qualifier Q = glm::defaultp>
+struct glmtype {
+    using type = glm::mat<C, R, T, Q>;
+};
+
+template <typename T, glm::qualifier P>
+struct glmtype<T, 1, 1, P> {
+    typedef T type;
+};
+
+template <typename T, glm::length_t L, glm::qualifier P>
+struct glmtype<T, L, 1, P> {
+    using type = glm::vec<L, T, P>;
+};
+
+template <typename T = double, glm::length_t C = 1, glm::length_t R = 1,
+          glm::qualifier Q = glm::defaultp>
+using glmtype_t = typename glmtype<T, C, R, Q>::type;
+
 }  // namespace util
+
+template <unsigned int Dim, typename Type>
+using Matrix = typename util::glmtype<Type, Dim, Dim>::type;
+
+template <unsigned int Dim, typename Type>
+using Vector = typename util::glmtype<Type, Dim, 1>::type;
 
 }  // namespace inviwo
