+++ include/inviwo/core/util/introspection.h
@@ -30,7 +30,7 @@
 #pragma once
 
 #include <inviwo/core/common/inviwocoredefine.h>
-#include <inviwo/core/util/glm.h>
+#include <inviwo/core/util/glmvec.h>
 #include <inviwo/core/util/document.h>
 
 #include <inviwo/core/util/detected.h>
